+++ src/client/java/dev/spiritstudios/snapper/mixin/PauseScreenMixin.java
@@ -6,10 +6,12 @@ import com.llamalad7.mixinextras.sugar.Local;
 import dev.spiritstudios.snapper.Snapper;
 import dev.spiritstudios.snapper.SnapperConfig;
 import dev.spiritstudios.snapper.gui.screen.GalleryScreen;
+import net.minecraft.client.Screenshot;
 import net.minecraft.client.gui.components.SpriteIconButton;
 import net.minecraft.client.gui.layouts.LinearLayout;
 import net.minecraft.client.gui.screens.PauseScreen;
 import net.minecraft.client.gui.screens.Screen;
+import net.minecraft.client.resources.sounds.SimpleSoundInstance;
 import net.minecraft.network.chat.Component;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.injection.At;
@@ -35,6 +37,22 @@ public abstract class PauseScreenMixin extends Screen {
             iconButtonRow.addChild(Snapper.createSnapperButton(20, _ -> {
                 this.minecraft.gui.setScreen(new GalleryScreen(this));
             }));
+        }
+        if (SnapperConfig.HOLDER.get().showScreenshotHelper()) {
+            iconButtonRow.addChild(SpriteIconButton.builder(
+                            Component.translatable("button.snapper.helper.screenshot"),
+                            _ -> {
+                                this.minecraft.getSoundManager().play(SimpleSoundInstance.forUI(Snapper.SHUTTER, 1.0F));
+                                this.minecraft.gui.setScreen(null);
+                                Screenshot.grab(this.minecraft, false);
+                            },
+                            true
+                    )
+                    .width(20)
+                    .sprite(Snapper.id("screenshots/helper"), 15, 15)
+                    .tooltip(Component.translatable("button.snapper.helper.screenshot"))
+                    .narration(_ -> Component.translatable("button.snapper.helper.screenshot"))
+                    .build());
         }
     }
 }